Jobs/Commands for checking/deleting floating attachments and command for deleting old deleted emails (deleted more than 3 months ago).

+++ app/Http/Controllers/Api/Email/EmailController.php
@@ -357,8 +357,11 @@ class EmailController
 
     public function deleteEmailAttachment(EmailAttachment $emailAttachment)
     {
-        //delete real file
-        Storage::disk('mail_attachments')->delete($emailAttachment->filename);
+        //delete real file (only when count on filename is 1, otherwise this attachment is also in use in another email because of a reply or send through)
+        $countAttachment = EmailAttachment::where('filename', $emailAttachment->filename)->count();
+        if($countAttachment == 1){
+            Storage::disk('mail_attachments')->delete($emailAttachment->filename);
+        }
 
         //delete db record
         $emailAttachment->delete();
